requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.3
opencv-python==4.4.0.44
Pillow==8.0.1
Werkzeug==1.0.1
